+++ triple_arena/src/nonzero_inx_vec.rs
@@ -344,6 +344,40 @@ impl<T> NonZeroInxVec<T> {
     }
     */
 
+    /// Shrinks the capacity of `self` as much as possible. The allocator may
+    /// give extra space for more capacity.
+    pub fn shrink_to_fit(&mut self) {
+        let new_cap = self.len();
+
+        if new_cap == self.capacity() {
+            // Don't do anything
+            return
+        }
+
+        if new_cap == 0 {
+            // this is all we have to do, capacity is set to 0 by this function
+            self.clear_and_shrink();
+            return
+        }
+
+        let new_layout = Layout::array::<T>(new_cap).unwrap();
+
+        // Safety: layout is for a nonzero sized allocation, nulls from the allocator
+        // are handled, we copy over correctly, we deallocate, and we set the
+        // new allocation pointer and new capacity
+        unsafe {
+            let new_ptr = alloc::alloc(new_layout);
+            if new_ptr.is_null() {
+                alloc::handle_alloc_error(new_layout)
+            }
+            let new_ptr = new_ptr as *mut T;
+            ptr::copy(self.allocation_ptr(), new_ptr, self.len());
+            alloc::dealloc(self.allocation_ptr() as *mut u8, self.layout());
+            self.set_allocation_ptr(new_ptr);
+            self.set_capacity(new_cap)
+        }
+    }
+
     #[inline]
     pub const fn nziter(&self) -> IntoNonZeroUsizeIterator {
         nzusize_iter(NZONE, self.len())