+++ triple_arena/CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
 ## [0.11.0] - TODO
+### Fixes
+- Fixed that the `Debug` impl of `PtrNoGen` was messed up
+
 ### Changes
 - Undid the tuple grouping of key value pairs in `OrdArena`, because it wasn't worth it and will not
   be compatible with the future cache local version anyway.
 - Renamed the `iterators` module to `arena_iterators`
+
+### Additions
+- Added `compress_and_shrink` and `compress_and_shrink_with` functions to all arenas
+- Added `<_ as Ptr>::name()`
 
 ## [0.10.0] - 2023-08-05
 ### Fixes